Add onmetal-kubeconfig flag
Introducing a new onmetal-kubeconfig flag to create the onmetal client.
Instead of using the credentials in the user-data secret, we will now
rely on a file passed to the MCM controller containing the kubeconfig.
The kubeconfig should be constructed in a way that the context contains
the desired target namespace which will be used for the onmetal resource
creation.